refactor: Car와 Cars 클래스를 car 패키지로 이동
