+++ src/test/java/racingcar/domain/generator/OilGeneratorTest.java
@@ -1,10 +1,9 @@
-package racingcar.domain;
+package racingcar.domain.generator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.generator.OilGenerator;
 import racingcar.exception.OilOutOfRangeException;
 
 import java.util.stream.IntStream;
